requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
anaconda-anon-usage
archspec
asgiref==3.8.1
attrs==24.2.0
backoff==2.2.1
boltons
branca==0.7.2
Brotli
certifi
cffi
charset-normalizer
click==8.1.7
colorama==0.4.6
conda
conda-content-trust
conda-libmamba-solver
conda-package-handling
conda_package_streaming
cryptography
decorator==5.1.1
distlib==0.3.8
distro
Django==5.1.1
django-cors-headers==4.4.0
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.16.1
folium==0.17.0
frozendict
frozenlist==1.4.1
future==1.0.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
gpsd-py3==0.3.0
idna
Jinja2==3.1.4
jsonpatch
jsonpointer==2.1
libmambapy
MarkupSafe==2.1.5
menuinst
mkl-fft
mkl-random
mkl-service==2.4.0
multidict==6.1.0
numpy
opencage==3.0.2
opencv-python==4.10.0
opencv-python-headless==4.10.0
packaging
pandas==2.2.3
phonenumbers==8.13.46
pillow==10.4.0
pipenv==2024.0.3
platformdirs
pluggy
pycosat
pycparser
pyfiglet==1.0.2
pyperclip==1.9.0
PySocks
python-dateutil==2.9.0.post0
pytz==2024.2
ratelim==0.1.6
requests
ruamel.yaml
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
tqdm
truststore
tzdata==2024.2
urllib3
virtualenv==20.26.6
wheel==0.43.0
xyzservices==2024.9.0
yarl==1.12.1
zstandard
